Move cache out of config
