+++ lyra-ui/src/apps.rs
@@ -1,11 +1,13 @@
 pub mod app_convert;
+mod appcache;
 mod applookup;
 
-use crate::config::{AppCache, AppsConfig};
+use crate::config::AppsConfig;
 use crate::plugin::{
   AppState, FuzzyMatchItem, OkAction, Plugin, PluginV, PluginValue, Renderable, SearchBlocker,
 };
 use anyhow::{anyhow, Context};
+use appcache::AppCache;
 use applookup::AppLookup;
 use egui::{Image, RichText};
 use lyra_common::convert as lyra_convert;
@@ -48,7 +50,7 @@ impl SearchBlocker for AppLaunch {}
 
 impl AppsPlugin {
   pub fn init(cfg: AppsConfig, cache_dir: &PathBuf) -> Result<Self, anyhow::Error> {
-    let cache = AppCache::load(cache_dir.join(format!("app_icons.toml")))?;
+    let cache = AppCache::load(cache_dir.join(format!("apps_icons.toml")))?;
     let apps = AppLookup {
       config: cfg.clone(),
       cache: Arc::new(cache),
